+++ receiver.go
@@ -39,22 +39,23 @@ type Receiver struct {
 	more                  bool                // if true, buf contains a partial message
 	msg                   Message             // current message being decoded
 
-	batching       bool                    // enable batching of message dispositions
-	batchMaxAge    time.Duration           // maximum time between the start n batch and sending the batch to the server
-	dispositions   chan messageDisposition // message dispositions are sent on this channel when batching is enabled
-	maxCredit      uint32                  // maximum allowed inflight messages
-	inFlight       inFlight                // used to track message disposition when rcv-settle-mode == second
-	manualCreditor *manualCreditor         // allows for credits to be managed manually (via calls to IssueCredit/DrainCredit)
+	autoSendFlow bool                    // automatically send flow frames as credit becomes available
+	batching     bool                    // enable batching of message dispositions
+	batchMaxAge  time.Duration           // maximum time between the start n batch and sending the batch to the server
+	dispositions chan messageDisposition // message dispositions are sent on this channel when batching is enabled
+	maxCredit    uint32                  // maximum allowed inflight messages
+	inFlight     inFlight                // used to track message disposition when rcv-settle-mode == second
+	creditor     creditor                // manages credits via calls to IssueCredit/DrainCredit
 }
 
 // IssueCredit adds credits to be requested in the next flow
 // request.
 func (r *Receiver) IssueCredit(credit uint32) error {
-	if r.manualCreditor == nil {
+	if r.autoSendFlow {
 		return errors.New("issueCredit can only be used with receiver links using manual credit management")
 	}
 
-	if err := r.manualCreditor.IssueCredit(credit); err != nil {
+	if err := r.creditor.IssueCredit(credit, r); err != nil {
 		return err
 	}
 
@@ -70,7 +71,7 @@ func (r *Receiver) IssueCredit(credit uint32) error {
 // DrainCredit sets the drain flag on the next flow frame and
 // waits for the drain to be acknowledged.
 func (r *Receiver) DrainCredit(ctx context.Context) error {
-	if r.manualCreditor == nil {
+	if r.autoSendFlow {
 		return errors.New("drain can only be used with receiver links using manual credit management")
 	}
 
@@ -80,7 +81,7 @@ func (r *Receiver) DrainCredit(ctx context.Context) error {
 	default:
 	}
 
-	return r.manualCreditor.Drain(ctx, r)
+	return r.creditor.Drain(ctx, r)
 }
 
 // Prefetched returns the next message that is stored in the Receiver's
@@ -406,6 +407,7 @@ func newReceiver(source string, session *Session, opts *ReceiverOptions) (*Recei
 			source:   &frames.Source{Address: source},
 			target:   new(frames.Target),
 		},
+		autoSendFlow:  true,
 		receiverReady: make(chan struct{}, 1),
 		batching:      defaultLinkBatching,
 		batchMaxAge:   defaultLinkBatchMaxAge,
@@ -448,7 +450,7 @@ func newReceiver(source string, session *Session, opts *ReceiverOptions) (*Recei
 		}
 	}
 	if opts.ManualCredits {
-		r.manualCreditor = &manualCreditor{}
+		r.autoSendFlow = false
 	}
 	if opts.MaxMessageSize > 0 {
 		r.l.maxMessageSize = opts.MaxMessageSize
@@ -496,13 +498,8 @@ func newReceiver(source string, session *Session, opts *ReceiverOptions) (*Recei
 // attach sends the Attach performative to establish the link with its parent session.
 // this is automatically called by the new*Link constructors.
 func (r *Receiver) attach(ctx context.Context) error {
-	// buffer rx to linkCredit so that conn.mux won't block
-	// attempting to send to a slow reader
-	if r.manualCreditor != nil {
-		r.l.rx = make(chan frames.FrameBody, r.maxCredit)
-	} else {
-		r.l.rx = make(chan frames.FrameBody, r.l.availableCredit)
-	}
+	// TODO: remove double-buffering
+	r.l.rx = make(chan frames.FrameBody, r.maxCredit)
 
 	if err := r.l.attach(ctx, func(pa *frames.PerformAttach) {
 		pa.Role = encoding.RoleReceiver
@@ -541,36 +538,36 @@ func (r *Receiver) mux() {
 		// unblock any in flight message dispositions
 		r.inFlight.clear(r.l.err)
 
-		// unblock any pending drain requests
-		if r.manualCreditor != nil {
-			r.manualCreditor.EndDrain()
+		if !r.autoSendFlow {
+			// unblock any pending drain requests
+			r.creditor.EndDrain()
 		}
 	}, func(fr frames.PerformTransfer) {
 		_ = r.muxReceive(fr)
 	})
 
 	for {
-		switch {
-		case r.manualCreditor != nil:
-			drain, credits := r.manualCreditor.FlowBits(r.l.availableCredit)
+		// max - (availableCredit + countUnsettled) == pending credit (i.e. credit we can reclaim)
+		// once we have pending credit equal to or greater than half our max, reclaim it.  we do this
+		// instead of pending > 0 to prevent flow frames from being too chatty.
+		if pendingCredit := r.maxCredit - (r.l.availableCredit + uint32(r.countUnsettled())); pendingCredit >= r.maxCredit/2 && r.autoSendFlow {
+			debug.Log(1, "receiver (auto): source: %s, inflight: %d, credit: %d, deliveryCount: %d, messages: %d, unsettled: %d, maxCredit: %d, settleMode: %s", r.l.source.Address, r.inFlight.len(), r.l.availableCredit, r.l.deliveryCount, len(r.messages), r.countUnsettled(), r.maxCredit, r.l.receiverSettleMode.String())
+			r.l.err = r.creditor.IssueCredit(pendingCredit, r)
+		} else if r.l.availableCredit == 0 {
+			debug.Log(1, "receiver (pause): source: %s, inflight: %d, credit: %d, deliveryCount: %d, messages: %d, unsettled: %d, maxCredit: %d, settleMode: %s", r.l.source.Address, r.inFlight.len(), r.l.availableCredit, r.l.deliveryCount, len(r.messages), r.countUnsettled(), r.maxCredit, r.l.receiverSettleMode.String())
+		}
 
-			if drain || credits > 0 {
-				debug.Log(1, "receiver (manual): source: %s, inflight: %d, credit: %d, creditsToAdd: %d, drain: %v, deliveryCount: %d, messages: %d, unsettled: %d, maxCredit : %d, settleMode: %s",
-					r.l.source.Address, r.inFlight.len(), r.l.availableCredit, credits, drain, r.l.deliveryCount, len(r.messages), r.countUnsettled(), r.maxCredit, r.l.receiverSettleMode.String())
+		if r.l.err != nil {
+			return
+		}
 
-				// send a flow frame.
-				r.l.err = r.muxFlow(credits, drain)
-			}
+		drain, credits := r.creditor.FlowBits(r.l.availableCredit)
+		if drain || credits > 0 {
+			debug.Log(1, "receiver (flow): source: %s, inflight: %d, credit: %d, creditsToAdd: %d, drain: %v, deliveryCount: %d, messages: %d, unsettled: %d, maxCredit: %d, settleMode: %s",
+				r.l.source.Address, r.inFlight.len(), r.l.availableCredit, credits, drain, r.l.deliveryCount, len(r.messages), r.countUnsettled(), r.maxCredit, r.l.receiverSettleMode.String())
 
-		// if receiver && half maxCredits have been processed, send more credits
-		case r.l.availableCredit+uint32(r.countUnsettled()) <= r.maxCredit/2:
-			debug.Log(1, "receiver (half): source: %s, inflight: %d, credit: %d, deliveryCount: %d, messages: %d, unsettled: %d, maxCredit : %d, settleMode: %s", r.l.source.Address, r.inFlight.len(), r.l.availableCredit, r.l.deliveryCount, len(r.messages), r.countUnsettled(), r.maxCredit, r.l.receiverSettleMode.String())
-
-			linkCredit := r.maxCredit - uint32(r.countUnsettled())
-			r.l.err = r.muxFlow(linkCredit, false)
-
-		case r.l.availableCredit == 0:
-			debug.Log(1, "receiver (pause): inflight: %d, credit: %d, deliveryCount: %d, messages: %d, unsettled: %d, maxCredit : %d, settleMode: %s", r.inFlight.len(), r.l.availableCredit, r.l.deliveryCount, len(r.messages), r.countUnsettled(), r.maxCredit, r.l.receiverSettleMode.String())
+			// send a flow frame.
+			r.l.err = r.muxFlow(credits, drain)
 		}
 
 		if r.l.err != nil {
@@ -656,9 +653,9 @@ func (r *Receiver) muxHandleFrame(fr frames.FrameBody) error {
 		if !fr.Echo {
 			// if the 'drain' flag has been set in the frame sent to the _receiver_ then
 			// we signal whomever is waiting (the service has seen and acknowledged our drain)
-			if fr.Drain && r.manualCreditor != nil {
+			if fr.Drain && !r.autoSendFlow {
 				r.l.availableCredit = 0 // we have no active credits at this point.
-				r.manualCreditor.EndDrain()
+				r.creditor.EndDrain()
 			}
 			return nil
 		}